Updated website per 1/24 meeting
added maintenance alert bar, changed the product list to a two column design, the product buttons links to a new page with the list of products for now, until I can figure out accordion cards. Also made the styles more consistent such as buttons and links. This is not mobile ready, so there will be spacing issues if you adjust the window size.